pyproject.toml
[project]
name = "webhook-relay"
version = "0.1.0"
description = "Receive webhooks from any service, forward to Telegram"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
